Explain explicitly setting hardhat network
Not explicitly setting the Hardhat network in `nox hardhat console` caught me out because I had assumed Hardhat would by default connect to the `localhost` network. It was my mistake, but it was very difficult to debug and stopped me from progressing through this awesome guide.

Because I think the default of creating and using a new, ephemeral network is unintuitive, adding a note explaining it could help other followers of guide not make the same mistake.