Update correct version of get balance script
Using the current version throws the following error: `eth is not defined`. While inspecting there is no `eth` variable available in the scope of the console instance. According to the following section of the documentation https://web3js.readthedocs.io/en/v1.7.3/web3-eth.html#getbalance the correct way of getting balance is as proposed.