Added a section on deploying to testnets and mainnets
Hey there - while I understand why testing on the local blockchain is great, I think a mention of testnets is merited so developers can understand the different of one vs the other. I work at Alchemy and we see developers ask about this. We also see that it's a common question for our developers to understand how to deploy to mainnet - therefore, I went ahead and added a high-level section there as deploying to mainnet would be the natural next step. I didn't want to clog up the tutorial with excessive details so I kept it as an overview and linked to a HardHat tutorial that showed how to deploy to mainnet. Lmk of any questions, thanks for having awesome docs!